refactor(checker): relocate tuple-rest spread queries off the common quarantine
The arch guard caps `query_boundaries/common.rs` at 1920 lines (#8225, must
not grow). Move the `tuple_variable_rest_offset` / `tuple_slice_variable_rest_offset`
query-boundary wrappers from `common.rs` to the call-specific boundary
`query_boundaries/checkers/call.rs`, where the spread call-argument logic
already lives. No behavior change.
